Fix timer blocking issue
The current set-up of time.sleep(5) was causing the GUI to freeze while
it slept. Instead, I've replaced this with a variable and a thread.
So, when the variable is set to False, a timer runs alongside the main
program, so that there's no freezing occurring.

Users can still press buttons, but the camera will not move until the
5 seconds are over.